--- airbyte-cdk/python/airbyte_cdk/base_python/cdk/__init__.py
@@ -1,1 +1,0 @@
-# Initialize SDK Package
