--- airbyte-cdk/python/airbyte_cdk/base_python/integration.py
@@ -1,118 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2020 Airbyte
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-
-import json
-import os
-import pkgutil
-from abc import ABC, abstractmethod
-from collections import defaultdict
-from typing import Any, Dict, Iterable, Mapping, MutableMapping, Optional
-
-from airbyte_cdk.models import AirbyteCatalog, AirbyteConnectionStatus, AirbyteMessage, ConfiguredAirbyteCatalog, ConnectorSpecification
-
-from .logger import AirbyteLogger
-
-
-class AirbyteSpec(object):
-    @staticmethod
-    def from_file(file_name: str):
-        with open(file_name) as file:
-            spec_text = file.read()
-        return AirbyteSpec(spec_text)
-
-    def __init__(self, spec_string):
-        self.spec_string = spec_string
-
-
-class Integration(ABC):
-
-    # can be overridden to change an input config
-    def configure(self, config: Mapping[str, Any], temp_dir: str) -> Mapping[str, Any]:
-        """
-        Persist config in temporary directory to run the Source job
-        """
-        config_path = os.path.join(temp_dir, "config.json")
-        self.write_config(config, config_path)
-        return config
-
-    @staticmethod
-    def read_config(config_path: str) -> Mapping[str, Any]:
-        with open(config_path, "r") as file:
-            contents = file.read()
-        return json.loads(contents)
-
-    @staticmethod
-    def write_config(config: Mapping[str, Any], config_path: str):
-        with open(config_path, "w") as fh:
-            fh.write(json.dumps(config))
-
-    def spec(self, logger: AirbyteLogger) -> ConnectorSpecification:
-        """
-        Returns the spec for this integration. The spec is a JSON-Schema object describing the required configurations (e.g: username and password)
-        required to run this integration.
-        """
-        raw_spec: Optional[bytes] = pkgutil.get_data(self.__class__.__module__.split(".")[0], "spec.json")
-        if not raw_spec:
-            raise ValueError("Unable to find spec.json.")
-        return ConnectorSpecification.parse_obj(json.loads(raw_spec))
-
-    @abstractmethod
-    def check(self, logger: AirbyteLogger, config: Mapping[str, Any]) -> AirbyteConnectionStatus:
-        """
-        Tests if the input configuration can be used to successfully connect to the integration e.g: if a provided Stripe API token can be used to connect
-        to the Stripe API.
-        """
-
-
-class Source(Integration, ABC):
-    # can be overridden to change an input state
-    def read_state(self, state_path: str) -> Dict[str, Any]:
-        if state_path:
-            state_obj = json.loads(open(state_path, "r").read())
-        else:
-            state_obj = {}
-        state = defaultdict(dict, state_obj)
-        return state
-
-    # can be overridden to change an input catalog
-    def read_catalog(self, catalog_path: str) -> ConfiguredAirbyteCatalog:
-        return ConfiguredAirbyteCatalog.parse_obj(self.read_config(catalog_path))
-
-    @abstractmethod
-    def read(
-        self, logger: AirbyteLogger, config: Mapping[str, Any], catalog: ConfiguredAirbyteCatalog, state: MutableMapping[str, Any] = None
-    ) -> Iterable[AirbyteMessage]:
-        """
-        Returns a generator of the AirbyteMessages generated by reading the source with the given configuration, catalog, and state.
-        """
-
-    @abstractmethod
-    def discover(self, logger: AirbyteLogger, config: Mapping[str, Any]) -> AirbyteCatalog:
-        """
-        Returns an AirbyteCatalog representing the available streams and fields in this integration. For example, given valid credentials to a
-        Postgres database, returns an Airbyte catalog where each postgres table is a stream, and each table column is a field.
-        """
-
-
-class Destination(Integration):
-    pass  # TODO
