+++ airbyte-cdk/python/airbyte_cdk/destinations/destination.py
@@ -1,0 +1,5 @@
+from airbyte_cdk.connector import Connector
+
+
+class Destination(Connector):
+    pass  # TODO
