+++ airbyte-cdk/python/airbyte_cdk/sources/singer/singer_helpers.py
