+++ airbyte-cdk/python/airbyte_cdk/sources/streams/__init__.py
