+++ airbyte-cdk/python/airbyte_cdk/sources/streams/http/auth/__init__.py
