+++ airbyte-cdk/python/airbyte_cdk/sources/streams/http/auth/core.py
