+++ airbyte-cdk/python/airbyte_cdk/sources/utils/casing.py
