+++ airbyte-cdk/python/airbyte_cdk/sources/utils/catalog_helpers.py
