+++ airbyte-cdk/python/airbyte_cdk/sources/utils/schema_helpers.py
